Correctly handle errors that occur when rolling back transactions
Before, errors when rolling back weren't handled correctly.  They
were not wrapped in a Sequel::DatabaseError subclass, and more
importantly, if the error was due to a disconnect, the
disconnection was not detected.  This is probably a reasonably
common code path when using transactions, as if there is a
disconnection, the BEGIN query will raise an exception, which
will cause the ROLLBACK query to be sent, which will raise a
separate exception.

This adds another begin/rescue around the rollback.  It also adds
specs for error handling when issuing begin/rollback/commit.