Allow json and xml serializers to support :array option in class to_json method to serialize existing array of model instances
Before, if you wanted to serialize an existing array of model
instances, you had to basically hack an existing dataset to do
so.  This is because the standard to_json method does not
pass options given to Array#to_json to the members of the array.
ActiveSupport works around this issue by overriding all of the
to_json methods, but we can't make the same change here.

To work around this issue, have the to_json method just
recognize the :array option, and have it use that instead of
calling #all to get the array of instances.