Fix possible thread safety issues on non-GVL ruby implementions in association reflection caching code
I don't believe these thread safety issues can happen on MRI due to
the GVL, so while there are still changes on MRI, it doesn't attempt
to use a mutex to synchronize access.

Basically, in modern versions of JRuby, Hash#[]= is not safe to use
by multiple threads simultaneously.  This adds a cached_fetch private
method that handles synchronized access to a subhash in the
association reflection that handles mutable data.  This way, you don't
need to synchronize access to the entire association reflection (which
is mostly immutable), just the parts that are mutable.  Note that
mutable here just refers to things that are cached on first call. I
expect that it is extremely difficult to hit any of the issues that
this commit fixes even in JRuby.