Make *_to_one association setters be noops if you pass a value that is the same as the cached value
If the object is already associated, there's generally no reason
to run the setter code.  Doing so can cause one_to_many cached
reciprocal associations to be reordered, and in the one_to_one
case, a couple pointless queries to be sent.

This can break code that manually monkeys with the associations
cache, which the nested_attributes plugin does, so add a
set_associated_object_if_same? method that determines whether
to run the association code anyway, and have the nested_attributes
plugin set that.

While here, refactor the one_to_one and many_to_setter setter
internals to call the same method, since the implementations
were almost identical.