Add :connection_handling=>:queue option to the threaded connection pools, may reduce chance of stale connections (Fixes #481)
Historically, Sequel has used a stack instead of a queue for
the threaded connections, since Array#pop/push is faster than
Array#pop/unshift in ruby.  However, that increases the chances
of connections getting stale, which can matter in some cases.

I'm currently keeping the default behavior the same, but it's
possible the default will change in a future version of Sequel,
so set :connection_handling=>:stack if you have a specific
desire to use a stack.