Make Database#transaction on PostgreSQL accept a :synchronous option to change synchronous_commit setting for that transaction
The synchronous_commit setting is per transaction and allows granular
control for determining when the transaction is considered committed.
The default setting waits until it is stored on disk on both the
current database and any slave databases.  You can set it to
:off or false to return before the transaction has been safely
committed to disk, which improves performance at the expense of
durability.  Newer versions of PostgreSQL also support :local and
:remote_write, read the PostgreSQL documentation for more details
on those settings.