Support :each when preparing statements, useful for iterating over large datasets
Before, only :select and :all were supported, which load the entire
dataset in memory.  On adapters that can use Dataset#each without
loading the entire result set in memory, using :each for pure
iteration should result in better performance.

While here, I discovered a bug that Database#call did not pass the
block it was given.  This could cause unexpected behavior, since
the block would not be called but results would still be returned.
That's because for other prepared statement types, a block is
optional.  For each it is required, which is how I discovered the
bug.