Add pg_json extension, for working with PostgreSQL 9.2's new json type
This is modeled on the pg_hstore and pg_array extensions, using
the same DelegateClass approach.

It does not currently support plain numbers or strings in a json
column, which PostgreSQL supports.  Support for that could
be added later if there is demand, just be adding custom parsers
and more DelegateClasses.  However, I think PostgreSQL is wrong
in the this case and it shouldn't be accepting those values,
which is why the extension doesn't currently support it.  Another
reason is that typecasting from a string would then be ambiguous,
since you wouldn't know whether it was literal JSON that should
be parsed, or represented a plain string that should be wrapped
in a DelegateClass.