Always attempt to load the ActiveSupport inflector in the plugin specs
This attempts to load the ActiveSupport inflector first, before
the Sequel inflector, since the Sequel inflector should then
override the ActiveSupport inflector.  Remove the check for the
ActiveSupport inflector in the inflector spec, since it should
never take affect.

You'll know if the ActiveSupport inflector is still in place if
the inflector specs start failing.