Make json_serializer plugin respect :root=>:collection and :root=>:instance options
The current behavior of the json_serializer plugin when the :root
option is given to the class or dataset to_json method is a bit
redundant, since it wraps both the collection and instances in
hashes.  I think changing the behavior is too risky in terms of
backwards compatibility, so I added additional options so that
users can pick the behavior they want.  I expect most users
will want :root=>:collection, and at some point in the future
that will probably become the default.