Make Dataset#get work with nil and false arguments
Before, if you passed nil or false, it ignored the argument and
used the block.  And if you didn't provide a block, it acted like
select_all.  Change it to use a default argument that can tell
whether an argument was provided.

This could potentially affect backwards compatibility, but
I think the previous behavior was broken.