Remove Database#case_sensitive_like on SQLite
Turns out that SQLite only has a setter PRAGMA for case sensitive
like, not a getter PRAGMA, so the previous getter always returned
false.

Note that the setter method is still there, so you can still
change the setting, you just can't easily check for whether the
setting is enabled or not.