Fix bug when parsing foreign keys for tables with explicit schema on PostgreSQL
There were multiple bugs here.  The first bug is that because a mutation
method was used for the join, but not for the where/select_more, the
where/select_more was skipped.  The second bug is that the where code
used an implicit alias instead of an implicit qualifier. The third
bug was that when the schema was used, an error would be raised
because it would try to call the output identifier method with
a qualified identifier.

Fix these bugs and add a spec for the behavior.