Nested attributes plugin: support creating children with primary keys passed through the attributes hash.
This adds an option to the nested_attributes declaration: :unmatched_key. Set it to :create to allow the creation of new children when an attributes hash contains an unmatched primary key. The default setting is :raise, which throws an error when an unmatched primary key is supplied. You can also set it to :ignore, which ignores any attributes hashes with unmatched primary keys. Setting :strict=>false is the same as :unmatched_pk=>:ignore, for backwards compatibility.

One small backwards incompatibility: when :unmatched_pk=>:ignore (or :strict=>false), the nested_attributes_update and nested_attributes_remove methods are no longer called when no children match the supplied primary key. Anyone overriding those methods should be aware of this.