Make Database#transaction on PostgreSQL recognize :read_only and :deferrable options
In addition to setting the transaction isolation level, PostgreSQL
also supports these additional options.  READ ONLY ensures that
most data changing commands are not allowed.  DEFERRABLE is only
important for transactions that are both SERIALIZABLE and READ
ONLY, in which case it has less overhead than a standard
SERIALIZABLE transaction.