Fix typecasting of time columns on jdbc/postgres, before could be off by a millisecond
I was getting this occassional failure on the jdbc/postgres tests,
which I tracked down to the floating point addition done here.
Instead of doing floating point addition after time parsing, put
the correct milliseconds as part of the string, so that the time
parsing code can get it correct without the need to worry about
rational numbers.