Add Database#global_index_namespace? method
This checks if the index namespace is global, or whether there is
a separate index namespace per table.  The default is true since
that is what most databases seem to use, with MySQL, MSSQL, and
Oracle being exceptions.

It's a shame this wasn't here in the beginning, since this
should be used by the default_index_name method to omit the
table name if indexes are already namespaced per table.  For
backward compatibility, that unfortunately cannot be done.