Namespace indexes when copying databases in some cases
If copying from a database without a global index namespace (e.g.
MySQL) to a database with a global index namespace (e.g. PostgreSQL),
the previous copying could break if you had indexes with the same
name in both tables.  To work around this, prepend the table name
to the index name when copying.

Note that if the index name exactly matches Sequel's default name,
it will not do any additional namespacing.