Fix jdbc adapter when DriverManager#getConnection fails, bump version to 3.36.1 (Fixes #488)
This is completely my fault, as the most basic code review should
have caught this.  It wasn't caught by the test suite since
DriverManager.getConnection does not fail on any of the jdbc
adapter tests in my environment (and I test on 10 separate jdbc
adapters).

Thanks to Alex Portnov for the heads up.