Fix offset-without-order code on Oracle
Offset without order is implemented by calling columns inside
of a select_sql call.  On oracle, if a limit is used, select_sql
is called with a clone of the dataset with append_sql already
set, which screws certain things up.  This makes sure the call
to columns doesn't affect the sql being generated.

While here, change from cspecify to specify in the spec, since
it's no longer checked.