Convert scalar types when typecasting Array to PGArray
Before, if an array object was passed to the type conversion
method, it was just wrapped in a PGArray.  Now in most cases
it will be converted using a recursive map and the appropriate
typecasting method.  In most cases it can pick the correct
method by default, but the :scalar_typecast option was added
so a specific method can be chosen.

For example, if you have an date[] field, you can pass an array
of strings (i.e. from a web form) to the setter, and it will
be typecasted to a PGArray of ruby Date objects.

This is implemented by having typecast_value_pg_array take an
optional third argument indicating the method to use for
typecasting scalars.

While here, refactor the tests so it tests conversions for
most array types that the extension supports.