Add pg_range extension, for dealing with PostgreSQL 9.2+ range types
Range types are a new feature in PostgreSQL 9.2.  They are similar
to ruby's Range class, but significantly more flexible.  Because
not all PostgreSQL Ranges can be converted to ruby Range objects,
the PGRange class is not a DelegateClass of Range, but a plain
subclass of Object.  However, in many cases it acts similarly to
a delegate class, as least if the PGRange can be represented by
a ruby Range.  Otherwise, this extension is similar to the other
Sequel pg_* extensions.