Add Database#extension and Dataset#extension for loading extension modules into objects automatically
This should make it easier to use some extensions.  Instead of
loading the extension with Sequel.extension and then manually
calling extend with full path of the extension module, you can
just call extension on the object with the extension name, and
it will require the extension and load the appropriate module.