Use change migrations instead of up/down migrations in the schema_dumper
change migrations are simpler, and the schema_dumper only used
create_table and add_index methods inside the up block, both of
which the MigrationReverser can reverse.  It's been over a year
and a half since reversible migrations have been added, so it
should be safe to make this switch.