Raise an Error if attempting to clone an association of differing type
I've done this a few times by accident when writing specs, and the
lack of an error makes it more difficult to track down the bug
in the specs.

I don't forsee problems with this, as the previous behavior was
unlikely to be used in working code.  I supposed it's possible for
one_to_many associations to be cloned to one_to_one or vice-versa,
since they share a lot of the same properties, but I don't know if
that's something we should support.