The json library is now required when running the plugin/extension specs
Previously, the extension specs tried to handle the lack of a json
library by skipping extensions or specs, similarly to how it handles
the lack of tzinfo or nokogiri.  However, because so many extensions
and plugins require json, and json is installed on almost all ruby
installations, I've decided to just require when running the specs,
which significantly simplifies the specs.