Switch to using :id_map instead of :key_hash
The :id_map option is much easier to understand, so remove most
of the documentation about :key_hash.  While here, remove the
description of the 3 argument form of the eager_loader proc.
That will remain supported indefinitely for backwards compatibility,
but I don't want to document it as it will encourage people to
use it.

While here, switch the identity_map and rcte_tree plugins to
use the :id_map option instead of the :key_hash option.