Allow for overriding the create/alter table generators used per Database object
This extracts a couple methods to make it easy for a Database
object to use a subclass of an existing generator class, allowing
for Database-specific methods to be added to generators in
shared adapters.

While here, rename the Generator class to CreateTableGenerator,
but keep Generator as an alias for backwards compatibility.