Do not automatically cast String to text in pg_auto_parameterize extension
The default type for literal strings on PostgreSQL is unknown, not
text, so casting to text is wrong and causes more problems than
it solves.

This breaks backwards compatibility for people who are using the
extension with values and expecting the automatic casting, but
it should make it significantly easier to switch to using the
extension for people that aren't.  If you were using this
extension and this causes a problem, you just need to explicitly
add .cast(String) to the string values.