Make filter by associations support use column references and method references correctly
When the column reference and method reference were split previously,
the regular loading and both types of eager loading were updated
to handle them correctly, but the filter by associations code was
not.  Make the changes to the filter by associations code so that
it handles column references and method references correctly.

In order for the filter by associations code to work with both
model instances and model datasets, you need to use different
parts of the reflection (method parts for the instance, column
parts for the dataset), so handle that correctly.

Add specs for this, and start testing for specific SQL in the
column reference versus method reference spec.  Also add
integration tests for filter by associations with keys in
PostgreSQL hstore columns.

While here, make the AssociationReflection#qualify method only
qualify Symbols and SQL::Identifiers, so it doesn't negatively
affect other expressions. A consequence of this is that
:predicate_key should never be needed to be set manually, you can
just set the regular option (such as :key), and if it is an
expression, it will work correctly.  Because it shouldn't be
necessarily to set :predicate_key manually, stop documenting it
and remove the mention from the CHANGELOG.

To make the implementation easier, setup some additional hash
entries when associations are defined.