Do deep qualification for non-column expressions in association keys
If a Sequel expression is given as an association column reference,
use the Qualifier AST Transformer to attempt to qualify internal
parts of it.

Without this, such expressions don't get qualified, which can lead
to issues later if joins are done manually.  Note that this does
not completely fix the issue, as the implicit qualification when
joining does not do deep qualification (yet).