Attempt to correctly qualify expressions when joining in model association datasets
This adds the :qualify=>:deep option to Dataset#join_table, which
reuses the Qualifier ASTTransformer subclass to qualify subexpressions
of expression objects given as keys/values to conditions.  Previously,
join_table just automatically qualified Symbols, now it can handle
many more types of objects.

Use this new feature in model associations, so you can provide an
unqualified expression as a column reference, and it will be qualified
correctly in all cases where it is used.  Update the specs to reflect
the new qualification. Also, add specs for filtering by associations
with model datasets.

Use this new feature to simplify the PostgreSQL hstore association
definitions, eliminating the manual qualification and the need for
:graph_only_conditions.

Because it's simple and some people may want it, also support a
:qualify=>false option in #join_table, which will skip all
qualification.