Allow bin/sequel to respect multiple -l options for logging to multiple files
Previously, the last -l option won.  The same behavior is now true
with the -E option, though specifying that multiple times just
gives you duplicate log entries on stdout.

The other change is that no options are passed when connecting to
the database.  Instead of specifying the :test and :logging as
database options, the loggers and testing of connnections is
set up after the Database object is instantiated.  This is mostly
for cleaner Database#inspect output.