Allow specifying eager_graph alias base on a per-call basis using an AliasedExpression
This was requested a while back, and is useful in limited
circumstances.  For example, let's say you had a Game class
with winner_id and loser_id fields that both point to the teams
table, and also a Player class with a team_id field that points
to the teams table.  Let's say you want to eager_graph both
the teams and their players for a set of games, and filter it
by some criteria on the players names.  Currently, you have
to do something like:

  ds = Game.eager_graph(:winner=>:players, :loser=>:players).
        filter(:players__name=>'A', :players_0__name=>'B')

The issue here is that the players table is going to be joined
twice in the query, once unaliased and once aliased to players_0.
In ruby 1.9, the winning players would be unaliased and the losing
players would aliased to players_0, but on ruby 1.8, it could be
reversed because hashes are unordered.  Even if you know that
players refers to the winning players and players_0 refers to the
losing players, it's still not very friendly and makes the filter
condition non-obvious.

There wasn't previously a good solution for this, but with this
commit, you can do the following:

  ds = Game.eager_graph(:winner=>Sequel.as(:players,:winning_players),
                        :loser=>Sequel.as(:players,:losing_players)).
        filter(:winning_players__name=>'A',
               :losing_players__name=>'B')