Remove core_extensions use from the specs
This moves testing of the core_extensions extension to a separate
rake task, and makes the remaining specs run without the core
extensions.

The idea here is that no internal part of Sequel should depend
on the core extensions.  Sequel previously had a spec that scanned
the source to look for possible core extensions use, but that is
not a fool proof approach.  If specs run clean without the core
extensions, it's unlikely anything internal uses the core extensions.