Make Sequel.expr(:symbol) split symbols
This makes Sequel.expr with a symbol return either an
SQL::Identifier, SQL::Qualified Identifier, or
SQL::AliasedExpression depending on the content of
the symbol.  Previously, it just wrapped the symbol in
an SQL::Wrapper, which made introspection more difficult.

This is slightly backwards incompatible, as some methods
are defined on SQL::Wrapper that are not defined on
SQL::AliasedExpression, such as #desc.

Doing this requires moving the split_symbol method out
of Sequel::Dataset and into the Sequel module, where it
is now a public singleton method.  The Dataset method now
just delegates to the singleton method.

With this change made, the hack used in Dataset#unaliased_expression
can be removed, as it isn't needed any longer.