Try to handle nil prepared statement values better on JDBC
On JDBC, the default way Sequel sets nil prepared statement values
does not work on all adapters in all cases.  That's because Sequel
generally does not know the type of the value being inserted, it
only knows that it is nil.  So it doesn't know the difference
between a nil value being inserted into an integer column, and a
nil value being inserted into a string column.

It turns out that JDBC has a reflection method to get the type
of the value.  However, this reflection method does not work on
all adapters (notably the jdbc/sqlserver adapter).  Use the
reflection method on derby and jtds, since it appears to fix
issues there.  While here, refactor and split off a method so
the overriding takes less overhead.