Attempt to handle blob values in the swift adapter
This will probably significantly slow down the swift adapter, but
as few people use it and this is more correct, I'm making the
change.  If anyone is using the swift adapter in production and
this causes too much of a performance hit, I'll open to making it
optional (similar to the jdbc adapter's convert_types setting).