Use real boolean literals on derby 10.7+ (Fixes #514)
The latest jdbc-derby gem is 10.6, and while Sequel doesn't run
perfectly on it, I try to maintain compatibility.  Real boolean
columns weren't added till derby 10.7.  Now, Sequel will check
for the svn version used, and if it indicates derby 10.7+,
TRUE and FALSE will be used instead of expressions that evaluate
to them.  This fixes using true/false as a column default.