Don't have pg_* extensions add methods to core classes unless the core_extensions extension is loaded
This adds methods to Sequel that act as replacements for the
methods added to Symbol, Array, Hash, and Range by the extensions.
If the core extensions are loaded, this is backwards compatible,
as the methods are still added.

The reasoning here is if you don't want Sequel's core_extensions,
you probably don't want to add methods to the core classes at all.