Don't allow model typecasting from string to postgres array, hstore, or composite types
User input for typecasting is very unlikely to come in a String
in PostgreSQL output format. The only reason a string was allowed
before was to allow the do/postgres, swift/postgres, and
jdbc/postgres adapters to work with the typecast_on_load
plugin.  Now that the pg_typecast_on_load plugin has been added,
this feature is no longer needed and should be removed.

The parsers for these types were written expecting it only to
be called with database output.  However, typecasting is called
with user input.  I don't know of any vulnerabilities in the
parsers, but as they weren't designed with security in mind,
it's better to protect them from user input.  At the very least,
the pg_array parser without sequel_pg is quite slow, so for large
input it's an easy DoS target.

Other types such as interval, inet, range, and json have much
simpler parsers that shouldn't pose any security issues, plus it's
conceivable that input could come in the form of a string for
those types.