Add :textsize option to tinytds adapter to override the default TEXTSIZE (Fixes #525)
The default TEXTSIZE setting on tinytds is about 64k, which is
pretty small.  While you can override the setting in freetds.conf,
for some users (Windows users using a binary gem), that might not
be the easiest change to make.  Allow a :textsize option to manual
set the setting for each connection.

While here, initialize the mssql unicode strings setting on
initialization instead of on connection.

Add a spec that tests >64k of data, and switch some test tables
to use generic types instead of specific types.