Add constraint_validations extension and plugin
The extension allows you to create database constraints that mirror
model validations when creating and altering tables.  It adds
validation metadata to a separate table.  The plugin reads the
metadata from that table, and uses it to automatically validate your
models before saving.

The advantage to doing so is that your validations are only specified
in one place, but are enforced both at the model level and at the
database level.  This is better for data integrity, since it
enforces the validation for all rows in the database, so changes
made by external programs or via the dataset layer are still
covered.