Recognize MySQL set type and default value (Fixes #529)
While here, refactor column_schema_to_ruby_default to not check
for database types.  Instead, override the method and call
super inside the shared adapters.

Instead of supporting set for all databases, only add it to the
shared mysql adapter.  Likewise, since interval is only supported
on PostgreSQL, move support for it to the shared postgres adapter.

For simplicity, still recognize the :set and :interval type symbols
in the global code in order to return the default values as a ruby
string, though that may be refactored out to the adapters later.

While here, do some minor cleaning up of the postgres adapter specs.