Add Postgres::PGRowOp#* for referencing the members of the composite type as separate columns
This acts differently than the #splat method, though the resulting
syntax is similar.  splat is for when you want to make sure you
are referencing a table as the table's composite type, usually
necessary when the table has the same name as one of its columns.
* is for referencing the members of the composite type as separate
columns.

The * method definition is copied from QualifyingMethods.
Unfortunately, the qualify method doesn't make sense for PGRowOp,
otherwise I would have just included the QualifyingMethods module.