Add :connection_handling=>:disconnect option for threaded connection pools
This option always disconnects the connection, instead of returning
it to the pool.  The expected use case is for people who want to
manually control all checkouts.

For example, let's say you have a web application that modifies
the connection state in each action.  You don't want a connection
used by one action to be used by another action, since you don't
know the state of the connection.  Also, later queries in the same
action may depend upon the state of the connection set previously
in the action, so you need to use a consistent connection for the
entire action.  In this case, a rack middleware that used
Database#synchronize around each action and used
:connection_handling=>:disconnect when connecting should work.