Add :where option to validates_unique, for custom uniqueness filters
validates_unique already takes a block, but the block is used for
scoping the existing uniqueness filter to a subset of the model's
dataset (for example, only validating uniqueness for active users
as opposed to all users).  This doesn't allow you to change the
actual filter used for uniqueness, which is necessary if you want
to do a case insensitive uniqueness validation for a database
that is case sensitive by default.

The :where option is used instead of the default filter, allowing
custom uniqueness checks.  In addition to case insensitive
uniqueness validations, it probably could be used for validations
of PostgreSQL exclusion constraints.