Allow subclasses plugin to take a block, which is called with all subclasses created
This is useful if you are creating subclasses at runtime.
Previously, if you wanted to make a specific change on a per
class basis to each subclass, you would have had to
periodically check descendents and look for new classes.
With this change, you can easily make such a change as soon
as the subclass is created.