Add bin/sequel -c support, for running code string instead of using an IRB prompt
You could previously have done this by piping the code to the
standard input, but this is definitely easier.  Would have
preferred to use the -e option so that executing code is similar
to ruby, but since that was taken, I choose to use -c since that
is similar to sh.