Refactor postgres adapter to work with new sequel_pg streaming
PostgreSQL removed the streaming support added in 9.2beta1, and
replaced it with a different API in 9.2beta3. The new API
uses separate PGresult objects for each row, so in those
cases, the connection execute methods return the connection
itself instead of a PGresult object.  Since the returned object
may not respond to clear, only clear it if it responds.

The _execute_prepared_statement method is being split off for
easier overriding.