No longer handle model classes as tables in join/graph methods in plain (non-model) datasets
Since 1.4.0, Sequel has checked the table argument when joining
and if it responded to table_name, it assumes it is a model
class and calls the table name argument on it.  I believe the
reason I choose to not directly depend on the Model class was
that I didn't want to couple the core and model layers.

Unfortunately, the current code has two flaws.  First, calling
respond_to? on arbitrary objects is not a great idea, since
other libraries may add methods with the same name to the
object.  For example, a String#table_name alias of #tableize.

Second, and more importantly, though model class's usually
represent the selection of all columns in a table, that is
not true all the time.  If a model's dataset isn't a simple
selection of all columns in a table, then joining just to the
table name is wrong, it should join to a subselect of the
model's dataset.

For that to work, it needs to recognize model classes and deal
with them.  Since core and model are decoupled, that cannot
happen at the core layer.  So the code to handle it moves to
the model layer.  There are two backwards incompatible side
effects of this change:

1) You can no longer use arbitrary objects that respond to
   table name.
2) You can only use model classes when joining to model
   datasets, you can no longer join a model class into a
   plain dataset.