No longer handle model instances in plain (non-model) datasets when inserting
Similar to the previous change, don't check respond_to and assume
the type.

When I took over maintenance of Sequel, Dataset#insert_sql had
a hard-coded check for Sequel::Model, which I replaced with a
respond_to check as the core should not depend on the model layer.

Unfortunately, values is not an uncommon method name, so it's
possible that other libraries can define the values method and
break Sequel.  This has already happened before, which required
checking that the result of values is a hash.

This moves the handling into Sequel::Model, and only handles
model instances.  Additionally, it contains a fix that allows
the pg_row plugin to work correctly.  Previously, if you provided
a model instance that used pg_row as the sole argument to
Dataset#insert, you would get an error, since it would use the
model's values as separate columns.

As before, the backward incompatible changes are you can no longer
just define values and you can only use model instances in model
datasets.