Make ValidationFailed and HookFailed exceptions have model method that returns the related model
Among other things, this makes it possible to use create and handle
errors when raise_on_save_failure is true:

  begin
    Model.create(:foo=>'bar')
  rescue ValidationFailed => e
    model = e.model
    # ...
  end

Before, you had to do:

  begin
    model = Model.new(:foo=>'bar')
    model.save

if you wanted to be able to reference the model if the saved failed.
Since saves can fail for both validations and hooks, extend this
feature to hooks as well.

While here, cleanup the exception message for hooks to explicitly
list which hook raised the error.