Allow overriding the convert_tinyint_to_bool setting on a per-Dataset basis in the mysql and mysql2 adapters
Unfortunately, because of the different APIs, the method used for
the per-Dataset overrides differs between the mysql and mysql2
adapters.  I was considering making the method name the same, but
on the mysql adapter, the method needs to take a field argument,
whereas on the mysql2 adapter, it doesn't, so the API would have
been different in arity anyway.

In the mysql adapter, you can override the cast_tinyint_integer?
method:

  def ds.cast_tinyint_integer?(field)
    super && true
  end

This can force the casting to integer even if the
convert_tinyint_to_bool setting is true.  Note that you cannot go
the other way currently, you can't force a casting to boolean
if the convert_tinyint_to_bool setting is false.  This is due to
how the conversion procs are handled internally.

The mysql adapter is flexible in that you can now cast tinyint(3)
columns to boolean if you want, instead of just tinyint(1)
columns.

In the mysql2 adapter, you can override the convert_tinyint_to_bool?
method:

  def ds.convert_tinyint_to_bool?
    true
  end

The mysql2 is flexible in that you can override both ways,
converting to boolean if the default is integer, and integer if the
default is boolean.

If you aren't sure which adapter you will be using, you can always
define both methods.

While here, enable to tinyint to boolean specs on the mysql2
adapter.  I guess it was an oversight that this wasn't already done.