Handle sets of alter_table operations on mysql and mssql where later operations depend on earlier ones
Some alter_table operations on mysql and mssql require parsing the
schema to get current metadata.  If you have multiple alter_table
operations that modify the same column, this can cause problems,
since all of the alter table queries are generated before being
run on the database.  For example:

  DB.alter_table(:foo)
    rename_column(:bar, :baz)
    set_column_type(:baz, Integer)
  end

On both MySQL and MSSQL, this would fail, since when generating
the set_column_type SQL, it would look for the baz column in the
schema, but that won't exist since the rename_column SQL has
not yet been run.

To fix this, preprocess the array of operations and split it
into multiple arrays.  Between the execution of SQL in each array,
remove the cached schema, so that the next set of operations will
have accurate schema information.

The implementation of the methods is the same for both of these,
except for one line (MSSQL doesn't need to parse the schema to
set a new default).  I thought about abstracting this into a
separate method, but I feel the implementations may diverge
further in the future, so I'm not convinced it is worth it.