Combine multiple alter_table operations into a single query where possible on MySQL and PostgreSQL
ActiveRecord has been doing this since 3.1, though just for MySQL
and not by default (it requires a separate :bulk=>true option).

This change may be risky, as it can change the SQL used for old
migrations, but hopefully it shouldn't cause problems.  This
preprocesses the alter table operations array and if it detects
operations that can't be combined (e.g. add/drop index operations
or rename column operations on PostgreSQL), it uses separate
queries for those.

If this causes problems that can't be easily fixed, this will be
turned off by default and only enabled via an option.  A simple
way to skip this combining manually is to use separate alter_table
blocks for each operation.

To implement this, alter_table_op_sql is split off
alter_table_sql, and returns just the fragment used.  Multiple
fragments can be separated by commas in a single ALTER TABLE query.

If the adapter does not support combining alter table operations,
then no changes are necessary.  However, if it does enable support
for combining operations, it must implement alter_table_op_sql to
handle overrides for the operations, or mark such operations as
not combinable.

It's possible to support something similar on Microsoft SQL
Server, but the syntax is different, and appears to be much
more limited, only adding multiple columns/constraints at once or
dropping multiple columns/constraints at once, but not mixing
adds and drops.  I don't have current plans to adding a similar
feature to Microsoft SQL Server.

Also, this changes the code to assume that alter_table_sql_list
returns an already flattened array.  Any custom adapters that
override alter_table_sql_list should make sure that they return a
flattened array.