Add Database#copy_into when using postgres adapter with pg driver, for very fast inserts into tables
This uses PostgreSQL's COPY table FROM STDIN support to get really
fast inserts when using data preformatted in CSV or PostgreSQL
text format.  Support for COPY table TO STDOUT was added last
year, but previous Sequel did not support that and you had to
drop down to the connection level, which doesn't offer an
API that is as nice.